refac: refactor read service about center posts for package dependency
